party_bot/requirements.txt
python-telegram-bot
openpyexcel
parse
